✅ Add "Async functions used in building test environments" to the tests
If the generation of a file or the execution of a command for testing fails, the failure must be identified.
However, Jest does not report the position where an asynchronous function that is not specified as an argument to the expect() function fails.
So I've updated the async function to run in the expect() function.