🎨 API tokens are required in order to access the GitHub API
The rate limit for the unauthenticated GitHub API is very low. Additionally, if a user overuses the unauthenticated GitHub API, they risk having their device's IP address blocked from GitHub.
The only situations in which users need to use the GitHub API are when running the CLI in a special environment where the git commands are not installed, and when the repository is private.
For this reason, we've changed it to require an access token in order to use the GitHub API.