chore(shipjs): do nothing with publishCommand when running "shipjs trigger"
This project will not be published to npm. Because of this, there is a `private` field in package.json and the `npm publish` command always fails.
When the `shipjs trigger` is executed, the `npm publish` command fails and the release process always crashes.

To fix this issue, I specified a command that does nothing in the `publishCommand` option of Ship.js.